♻️ Move permission check to permission screen
